+++ lib/timely/rails/period.rb
@@ -29,5 +29,30 @@ module Timely
     def to_s
       "#{number} #{units.to_s.gsub('_', '')}"
     end
+
+    def to_seconds
+      case units
+      when :seconds
+        number
+      when :minutes
+        number.minute.in_seconds
+      when :hours
+        number.hour.in_seconds
+      when :days
+        number.day.in_seconds
+      when :weeks
+        number.week.in_seconds
+      when :months
+        number.month.in_seconds
+      when :years
+        number.year.in_seconds
+      when :calendar_days
+        ::Time.now.advance(days: number - 1).end_of_day - ::Time.now
+      when :calendar_months
+        ::Time.now.advance(months: number - 1).end_of_month - ::Time.now
+      when :calendar_years
+        ::Time.now.advance(years: number - 1).end_of_year - ::Time.now
+      end
+    end
   end
 end