+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- [TF-36] Add to_seconds method in Period class
+
 ## 0.13.0
 
 - [PLAT-1175] Update to Ruby 3.2